Rename i2c module as 'devices'
It's going to be managing other kinds of devices too soon, so a more
generic name is appropriate.
